Move db open fixture to conftest
