+++ tests/unit_tests/model/test_mirrors.py
@@ -6,17 +6,10 @@ import pytest
 
 import simtools.config as cfg
 import simtools.io_handler as io
-from simtools import db_handler
 from simtools.model.mirrors import Mirrors
 
 logger = logging.getLogger()
 logger.setLevel(logging.DEBUG)
-
-
-@pytest.fixture
-def db(set_db):
-    db = db_handler.DatabaseHandler()
-    return db
 
 
 def test_read_list(db):
