+++ tests/unit_tests/model/test_telescope_model.py
@@ -8,17 +8,10 @@ import pytest
 
 import simtools.config as cfg
 import simtools.io_handler as io
-from simtools import db_handler
 from simtools.model.telescope_model import InvalidParameter, TelescopeModel
 
 logger = logging.getLogger()
 logger.setLevel(logging.DEBUG)
-
-
-@pytest.fixture
-def db(set_db):
-    db = db_handler.DatabaseHandler()
-    return db
 
 
 @pytest.fixture
